xml 어노테이션 변경

See merge request s12-final/S12P31D208!42

+++ fitmily_backend/user-service/src/main/java/com/d208/fitmily/domain/health/dto/HealthInsertDto.java
@@ -12,7 +12,7 @@ public class HealthInsertDto {
     private Integer userId;
     private Float height;
     private Float weight;
-    private Float bmi;
+    private Integer bmi;
 
     // JSON 문자열로 직렬화한 결과를 담을 필드
     private String healthOtherDiseasesJson;
